\name{item.summary}
\alias{item.summary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{create various levels of summary for an individual item}
\description{This service parses the metadata record and creates a summary specific to the theme, with subtype being defined by the attribute name}
\usage{
item.service()
}
%- maybe also 'usage' for other objects documented here.

\value{
A JSON object as summary holder
}
\author{
Jordan S. Read
}
\seealso{
\code{historical.service}
\code{storm.service}
\code{vulnerability.service}
}


\keyword{coastal}
\keyword{hazards}


